Update IOS and Android TextFields extend their respective Components and implement MobileTextField.
